MAM-3835-unable-to-create-new-mothsocial-account-if-language-set-to-brazilian (#371)
The Mastodon server will return an error it doesn't have an exact match for the localization language being requested in the signup flow. The app requests "pt" in this case, but the server only has "pt-BR" and "pt-PT"

For now, request that the confirmation email be sent in English, and longer term, look at doing a re-request if we get this error.